refactor database to singletons
